Retire NetUtils
The only reference to NetUtils is where !InetSetup attempts to load it (silently ignoring any fail).
It also outputs this command into "Internet.Startup" but after an RMEnsure of NetFS 5.79.
Since these are ROM components files, and in every case NetFS 5.79 is present, NetUtils is redundant. !InetSetup should continue as before since it is part of the universal boot sequence and may find itself on a machine with an older NetFS.
Also - moved RTCAdjust up into alphabetical order (this handily keeps the module numbering the same for Tungsten & IOMD as a side effect).

Version 6.05. Tagged as 'BuildSys-6_05'