Split library include paths into separate included Makefile
This de-duplicates some of the information from AppLibs and ModuleLibs,
which now include the new Makefile fragment in order to ensure compatibility.
It also means that the definitions can be used for building other library
components, which by definition do not perform a link step, without having
to choose either AppLibs or ModuleLibs when most of either one is unused for
a library build.