Allow components using CLibrary shared Makefile to specify additional exports
Occasionally, a library needs to export additional files that don't live in
either an h or hdr directory and (when cross compiling) shouldn't have a
.h suffix added to their exported version. A typical example would be where
the licence conditions of the library require a licence header to be attached
to all copies. To achieve this, CLibrary now uses EXPORTS in a similar manner
to how AAsmModule does, simply as a dependency of export_hdrs, which can
expand to any additional targets you need to define. By default, it overrides
the internal targets EXPORTING_HDRS and EXPORTING_ASMHDRS (derived from
HDRS and ASMHDRS supplied by the master makefile), but you can always include
them in your definition of EXPORTS if desired.