Add tokenised help file functionality to `CModule`
`TOKHELPSRC`, `HELPSRC` and `TOKENS` have the same meaning as in `AAsmModule`.
Because `CModule`, unlike `AAsmModule`, can build a binary from multiple
linked object files, it is also necessary to specify which object file(s)
depend on the autogenerated source file. This is achieved with the new input
variable `TOKHELPDEPENDS`. This defaults to `OBJS` which in turn defaults to
`TARGET`, which should mean that any components converted from `AAsmModule`
will not require any such line to be specified in their master makefile.

Requires RiscOS/Sources/Programmer/Debugger!4