Change a few modules to `C` type
These components have pending merge requests that mean they are likely to
move to use the `CModule` shared makefile in the near future, which will
require the use of their `rom_link` rule. However, due to the change to
`AAsmModule` in the last commit, we can change their type now in anticipation.

I'm not going further and change all `ASM` components en masse, because not
all of them will yet have been converted to use `AAsmModule`. This is also
the reason for not making the change in `srcbuild` itself.

Version 7.49. Tagged as 'BuildSys-7_49'