Support links for extension-less files in subdirectories
This only applies to cross-compilation builds. The `make links` rules don't
know a pre-determined list of subdirectories that need to be converted into
filename extensions. Instead, files that should have no filename extension
are usually simply identified by the absence of a `/` in their specification
in `SOURCES_TO_SYMLINK`. However, this is insufficient to handle some rarer
cases, such as a `Makefile` within a `Test` subdirectory.

To handle this, we introduce explicit support for specifying such files in
`SOURCES_TO_SYMLINK`, using a double slash where there would otherwise be a
filetype subdirectory name, such as `Test//Makefile`. (Conveniently, this is
also valid Posix filename syntax anyway.) The way this is achieved is to
tweak the regular expressions to require the directory and leafnames between
slashes to be at least one character long.

Version 7.58. Tagged as 'BuildSys-7_58'