Enable `CmdHelp` merging rule when `CUSTOMRES=no` and cross-compiling
In the cross-compiling case, a rule is defined targeting `MERGEDMSGS`, which
the phony target `resources_cmdhelp` depends on, but the definition is
conditional on `CUSTOMRES` not being `custom` or `no`.

In a use-case which is believed to be unique to LanManFS, it is required
that the resource files not be registered to ResourceFS via the Messages
module, but by LanManFS itself, both for RAM and ROM builds, but the merging
of `CmdHelp` and `Messages` files is still required. To leverage the
`Messages` merging code in `CModule`, re-enable the `MERGEDMSGS` rule when
`CUSTOMRES=no` (which is arguably the correct setting for LanManFS) but leave
it disabled for `CUSTOMRES=custom` to permit full customisation if required.