Stop CApp from assuming Messages is installed during 'resources'
By contrast with INSTRES_FILES in CModule, where Messages is implicitly
included, CApp and BasicApp do not do this for RES_FILES. If used with
a component that doesn't feature a Messages file this results in an error.
The native-compilation version of CApp was doing duplicate installations
of Messages, which was at odds with the cross-compiling CApp and both
versions of BasicApp.

The only existing CApp client components that define RES_FILES are
'Help2' and 'EditApp', and they both explicitly list Messages, so this
change should not break either of them.