Add initial support for HAL & kernel library builds
Extend the CLibrary shared makefile to add support for building HAL and
kernel versions of libraries, with 'h' and 'k' suffixes respectively.
These are built with stack limit checking & frame pointers disabled, and
with RISCOS_HAL and RISCOS_KERNEL defines enabled as appropriate.

Note that currently there's nothing done to protect the SB register in
HAL builds of libraries, or to deal with references to data sections.