Add HostApp makefile fragments
The HostApp fragment is designed to be used for building apps/tools for
running on the build host. It's currently fairly basic (just a
stripped-down and tweaked version of CApp), but it should be sufficient
for most purposes. The main advantage of this over HostTools+CApp (as
used previously) is that HostApp will make sure to use the host's
version of the CLib stubs, avoiding any incompatibilities between the
host and target CLib versions.

RISC OS version has been tested with a Disc build (after modifying the
relevant HostTools-using components), GNU version is untested.

Version 7.81. Tagged as 'BuildSys-7_81'