Change default RAMASMDEFINES to mirror AAsmModule
The AAsmModule shared makefile fragment predefines two assembly variables
by default when targeting a softload binary: `standalone` (set to {TRUE})
and `MergedMsgs` (set to the path to the merged `CmdHelp`/`Messages` file).
It's reasonably common for components to migrate over time from AAsmModule
to CModule; to assist with this process, change the default setting for
`RAMASMDEFINES` to match.

Once this is merged, the following components could have their Makefiles
simplified: Debugger, MakePSFont, NVidia and VFPSupport.

Version 7.82. Tagged as 'BuildSys-7_82'