Change how top-level phony rules are overridden
A common pattern previously in use was for a shared makefile to define a
top-level phony rule (the sort of thing that srcbuild or a !Mk file would
use as the target to pass to amu) but with a macro expansion within the
rule name. This allowed clients to override the rule by assigning any value
to the macro.

However, this approach makes it relatively difficult for clients to be
adapted to support cross-compilation. This is because when cross-compiling,
the shared makefiles utilise nested invocations of the make tool in order
to facilitate an out-of-tree build directory layout. Therefore, clients
need to determine whether they are cross-compiling, and if so, whether they
are in the context of the inner or outer make invocation, before they can
decide whether or not to redefine the phony target. This can be achieved in
as little as 3 extra lines:

  CURDIR ?= objs
  ifeq (objs,$(notdir ${CURDIR}))
  # redefinitions of top-level phony rules placed here
  endif

but this wasn't universally popular. Instead, taking inspiration from
INSTTYPE, we can chain from each ${PHONY} rule to a ${PHONY}_custom rule
and have the client makefile define the ${PHONY}_custom rule instead.
Crucially, in the cross-compiling case, the shared makefile can then take
care of ensuring this only happens on the inner make invocation.

In summary:

When setting `CUSTOMEXP` to `custom`, clients should henceforth
* define `export_hdrs_custom` and `export_libs_custom` rather than redefine
  `export`
* define `export_hdrs_custom` rather than redefine `export_hdrs`
* define `export_libs_custom` rather than redefine `export_libs`

When setting `CUSTOMROM` to `custom`, clients should henceforth
* define `install_rom_custom` rather than redefine `install_rom`
* define `rom_custom` rather than redefine `rom`
* define `rom_link_custom` rather than redefine `rom_link`

When setting `CUSTOMSA` to `custom`, clients should henceforth
* define `install_custom` rather than redefine `install`
* define `standalone_custom` rather than redefine `standalone`

`CUSTOMDBG` and `CUSTOMGPA` are now removed. These were not being used by
any client makefiles, but it seems probable that the rules they related to
(`debug` and `gpa_debug` respectively) would have needed overriding in
precisely the same circumstances as `install` and `standalone`, so in
future `CUSTOMSA` should be used for these also.

Version 7.91. Tagged as 'BuildSys-7_91'